net: use new needs_fwbr helper
pve-firewall introduced a new helper for deciding whether to create a
fwbr for a specific bridge. Switch over to the new helper to avoid
duplicating the decision logic across multiple repositories.

This also fixes an issue with OVS and the nftables firewall, where
firewall bridges are still required in order for the VM firewall to
work.

Removed the comment, since it is now inaccurate. For more information
look at the documentation of the new helper instead.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>