api: introduce migration preconditions api endpoint
Add a migration preconditions api endpoint for containers in similar
vein to the one which is already present for virtual machines.

This is needed to inform callees about positive and negative ha resource
affinity rules, which the container is part of. These inform callees
about any comigrated resources or blocking resources that are caused by
the resource affinity rules.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730181428.392906-20-d.kral@proxmox.com