api: migration checks: rename return prop to dependent-ha-resources and improve description
We actually query if there are any guests positive affinity rule for
the to-be migrated CT, while that normally means the they will be
migrated, it doesn't has to be (e.g., node constraints might interfere
here), and "comigrated" is not as much used compared to
"dependencies", so that might be easier to understand for non-native
speakers or users (vs devs, these details tend to leak).

We did not publish this to a stable release, so ignore the api
breakage.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>