setup: debian: increase upper supported version limit from <= 13 to <15
Using a less-than-or-equals comparison is rather bad, as we compare
the X.Y version here, e.g. 14.0 or 14.1, so the initial .0 release
will be accepted but the first point release suddenly won't, which
makes no sense and is much more subtle and much less expected for
everyone (devs and users) compared to the initial major version
getting rejected already, as then one can notice this explicitly.

So drop the "equal" part from the comparison and rather increase the
version used in the check by one to allow all (point) releases of a
specific major release.

This was initially reported on GH (see link below) and accompanied
with a pull request, as we cannot accept PRs on GitHub and as the PR
still use a smaller-than-or-equal comparison and due to the regression
potential of this I decided to create my own patch, but I'd like to
record my thanks for the reporter here.

Link: https://github.com/proxmox/pve-container/pull/2
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
(cherry picked from commit b69cdbc21fb070ed2f25fbec9a1c22aa1c76fd82)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>