setup: debian: regenerate snakeoil certificate and keypair
Debian-based container templates that have the 'ssl-cert' package
installed will contain a TLS certificate and keypair generated at
template creation time. it needs to be regenerated per-instance to
avoid accidental use of a known private key.

The logic here only triggers on newly created containers, not on
clones or restored backups, where deleting the key/certificate might
be the wrong thing to do.

Reported-by: Tijl Deneut <tijl.deneut@ebo-enterprises.com>
Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>