setup: do not throw a hard error if Debian version is higher than known maximum
As the version can increase after a user (which might not even have
access to the PVE system, did a system upgrade inside an existing
container it's really far from ideal to suddenly break the start.

As while it's normally good and warranted to ensure that one can
understand the version one tries to run, there is hardly any data loss
potential here, and with modern systemd based environments there is
much less distro specific that can suddenly break. So instead log a
prominent warning through our task worker warning infrastructure that
makes it visible in the task result.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>