api: create/store: allow adding CT as HA resource after creation
Extend the creation and restore actions with a 'ha-managed' parameter
that, if enabled, will also add the CT as a new HA resource. The 'state'
parameter for this new resource will match the value of the 'start'
parameter used during creation of the CT, such that the resulting state
of the resource and CT both match the user's expectation (avoid
situation where user creates a CT, does not select 'Start after
creation', but the default 'started' state of the resource would start
the CT anyway).

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20251006155233.267374-3-m.koeppl@proxmox.com