d/control: record dependency for isc-dhcp-client
For now the OCI implementation uses the dhclient, which is shipped by
isc-dhcp-client.

We will need to move away from that dhcp client implementation sooner
or later, as it's deprecated upstream [0] and will also vanish from
Debian repos sooner or later. But for Debian Trixie it's still there
so not a problem for now. Alternatives might be the busybox dhcp
client or the one from dhcpcd-base, but as we rely on the isc one in
other places and as this is mostly an implementation detail, we can
switch this later.

[0]: https://www.isc.org/blogs/dhcp-client-relay-eom/

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>