config: add env property for configuring the CT environment
This can be useful in general, but will be especially useful for the
new OCI containers, as there configuration often happens over these
environment variables.

We re-use the (relatively new) lxc.environment.runtime config keys and
map that to an NUL-separated string saved in a env property on config
read, this ensures that existing manually added environment
definitions do not get lost. On serializing we do the reverse, this
combination avoids the need for another array or list handling for the
config itself, and the one for the native lxc keys is already pretty
mature.

Currently there is only a smaller rough edge around editing an pending
env value to empty, but the more prominent revert works just fine, so
let this edge case be for now.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>