api: config: censor env property for users without VM.Config.Options
The reasoning here is that these can include credentials or other
information that one would like to not expose to users with just audit
permissions, as they might be used to take over a CT or the service it
provides. For now limit them to those whom can edit them, i.e. users
with VM.Config.Options, if the need arises for users to see these
options without having to give them that (relatively powerful)
privilege, we can always add a new (dedicated) one for providing
access to these options.

Note that as of now all lxc.* items won't be returned in anycase, so
this is not a limitation of the status quo, but really just affects
the recently (not yet bumped) "env" property.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>