config: replace global ipmanagehost with a netdev specific host-managed option
The global option was named a bit oddly and as the whole management
code was already prepared to handle this on a per-interface basis, it
just makes sense to move this into the network device config property.

That way we can also expose it nicely in the UI in the existing netdev
widgets and allow to use this for non-OCI CTs too (not really tested
currently though).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>