create: improve encapsulation for detecting and extracting OCI tar image
Avoid that the API endpoint has to care about details that we assume
for when detecting OCI images, rather encapsulate that in the
archive_is_oci_format and restore_oci_archive methods directly.

With that we can simplify the branches and logic in the create API
endpoint again.

This aligns the method's signature closer to what we use for similar
extract/restore ones.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>