oci create: ensure values taken from image do not contain control characters
These values end up in our configuration, so control characters would
not be great to have, as these effectively allow one to inject
arbitrary config keys, which in turn might allow "funny" stuff like
mapping some files from root in the ct.

It probably would be better to handle this in the rust code extracting
the OCI config from the image, but for now this here should do the
trick.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>