create oci: best-effort create directories of data volumes inside the CT
Ensure all volume paths mentioned in an OCI image config exist in the
target container root filesystem.

Note that this is not yet fully complete with what the spec recommends
(but does not enforces), as that would be:

> Implementations SHOULD provide mounts for these locations such that
> application data is not written to the container's root filesystem.
-- https://github.com/opencontainers/image-spec/blob/v1.1.1/conversion.md#optional-fields

But for us that isn't simple with the current approach. In the midterm
we would like to provide a new storage content type for a more native
OCI images representation (keeping the content addressable blobs
directly on the file system) and then leverage that for actual layered
filesystem with explicit data volumes configured by the user.
OTOH parsing the tar in sync during the create wizard to provide
the list of required data volumes is just rather overkill for now,
and would not even provide that much of a benefit with the current
approach (where the result is basically a closer to full blown system
CTs).

In any case, not all images provide the base directories paths for the
listed volumes in their filesystem trees, so  we need to create them
in any case. That fixes a few OCI applications, for example uptime
kuma.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>