config: clarify description of entrypoint property
The previous description implied that `entrypoint` must be an absolute
path to a binary within the container's rootfs. In reality, the
entrypoint may be any command, optionally with arguments. It can start
with an absolute path, a path relative to the container's working
directory, or the name of a binary found in the containers PATH
environment.

Signed-off-by: Filip Schauer <f.schauer@proxmox.com>
Link: https://lore.proxmox.com/20251118112556.51942-1-f.schauer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>