config: allow tabs for values of environment variables and entrypoint
Quite some dockerfiles (often the source of a OCI image) are written
such that some strings are wrapped over multiple lines and intended
with tabs, which in turn will turn up in the resulting OCI image
manifest variable. For us the \tab character is not really a problem
during de/serialization, so just allow it.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>