fix #6897: warn that nesting may be required for systemd
Recent versions of systemd require nesting to isolate services. If
nesting is disabled Debian 11 and 12 containers hang for 25 seconds
after login and Debian 13 just shows an empty console. To make this
less confusing for users, add a task-log warning on CT start if a
systemd version >241 (used by Debian 10) is detected.

Signed-off-by: Robert Obkircher <r.obkircher@proxmox.com>
Link: https://lore.proxmox.com/20251119142738.26840-4-r.obkircher@proxmox.com
[FE: style fix - no parentheses for post-if expression]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>