lxc config: mount /dev/shm tmpfs for application containers
Mount a tmpfs at /dev/shm into containers without an init system which
would normally set up a shared memory filesystem. This satisfies the OCI
Runtime Spec [0], which specifies that the following should be made
available in the container:
/proc       proc
/sys        sysfs
/dev/pts    devpts
/dev/shm    tmpfs

The reason the mount is added conditionally is to avoid interfering with
init systems: systemd correctly detects an existing /dev/shm, but OpenRC
mounts it even if already present, resulting in duplicate mounts.

[0] https://github.com/opencontainers/runtime-spec/blob/fd25dd34a06b/config-linux.md#default-filesystems

Signed-off-by: Filip Schauer <f.schauer@proxmox.com>
Reviewed-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Link: https://lore.proxmox.com/20251202152733.204787-1-f.schauer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>