fix #7156: setup: separate dns list with semicolons
Network Manager uses GLib's key file format [1]. While the format allows
either comma or semicolon-separated lists, this choice is global for the
key file and Network Manager declares that their key file uses the
later. As per `man 5 nm-settings-keyfile`:

> lists are separated by character ;

Additionally, tables 6 and 7 from the manual contain examples for dns:

> Example: dns=1.2.3.4;8.8.8.8;8.8.4.4;
> Example: dns=2001:4860:4860::8888;2001:4860:4860::8844;

There are, however, a handful of lists mentioned in the man page which
use a different format:

> some lists of complex values (addresses, routes, routing-rules),
> instead of using a semicolon separated list, use one key-value pair
> per list element, with the key being the singular of the property name
> followed by the numeric index (i.e address1, address2, ...).

so care should be taken when revisiting lists in this context. For
example, routes are not lists per se (they are strings) but they
allow (cf. table 6):

> Example: route1=8.8.8.0/24,10.1.1.1,77

[1] https://docs.gtk.org/glib/struct.KeyFile.html

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Link: https://lore.proxmox.com/20251218150126.832313-1-m.sandoval@proxmox.com